fix id type

+++ src/main/java/com/smartparking/parking_api/repository/VehiculeTypeRepository.java
@@ -3,5 +3,5 @@ package com.smartparking.parking_api.repository;
 import com.smartparking.parking_api.entity.VehiculeType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VehiculeTypeRepository extends JpaRepository<VehiculeType, Long> {
+public interface VehiculeTypeRepository extends JpaRepository<VehiculeType,Integer> {
 }
